Create NOT_related_script.js
This along with the index.html file where supposed to make up my JavaScript version of the OpenWeatherMap assignment.  I was able to successfully call the API and get a return of JSON data, but I only had luck when hard coding the API key into the query string.  I also had trouble getting the innerHTML property to work so the only evidence of successful output is in the console.log output to theJavaScript console.